requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2023.3
sqlparse==0.4.4
